refactor: update amount type to Integer and enhance error handling in NFC scan flow
- Changed the `amount` parameter type from `Double` to `Int` in `MemberRepository.addAmount`.
- Updated `NfcScanScreen` to convert the invoice amount input from `Double` to `Int` before processing.
- Refined error handling in `NfcScanScreen` to prevent falling back to incomplete member data when verification fails or is rejected.
- Added specific error messages ("Card Not Registered" and "Verification Rejected") and ensured `verifiedMemberId` is cleared upon verification failure.

+++ composeApp/src/commonMain/kotlin/com/example/sitacardent/feature/home/data/MemberRepository.kt
@@ -72,7 +72,7 @@ class MemberRepository {
         }
     }
 
-    suspend fun addAmount(memberId: String, amount: Double, cardMfid: String, password: String): Result<AddAmountResponse> {
+    suspend fun addAmount(memberId: String, amount: Int, cardMfid: String, password: String): Result<AddAmountResponse> {
         println("MemberRepository: Adding Amount - ID: $memberId, Amount: $amount, Card MFID: $cardMfid, Password: [PROTECTED]")
         return try {
             val response: AddAmountResponse = client.post("$baseUrl/add-amount") {
